+++ src/blocks/faq/index.js
@@ -12,8 +12,9 @@ import deprecated from './deprecated/';
 import edit from './edit';
 import metadata from './block.json';
 import save from './save';
+import transforms from './transforms';
 import { __ } from '@wordpress/i18n';
-import { createBlock } from '@wordpress/blocks';
+
 
 const { name } = metadata;
 
@@ -68,33 +69,5 @@ export const settings = {
 	save,
 	edit,
 	deprecated,
-	transforms: {
-		to: [
-			{
-				type: 'block',
-				blocks: ['vk-blocks/faq2'],
-				transform: (attributes, innerBlocks) => {
-					const questionInner = [
-						createBlock('core/paragraph', {
-							content: attributes.heading,
-						}),
-					];
-					const questionBlock = createBlock(
-						'vk-blocks/faq2-q',
-						{},
-						questionInner
-					);
-					const answerBlock = createBlock(
-						'vk-blocks/faq2-a',
-						{},
-						innerBlocks
-					);
-					return createBlock('vk-blocks/faq2', {}, [
-						questionBlock,
-						answerBlock,
-					]);
-				},
-			},
-		],
-	},
+	transforms,
 };
